terrain movement modifiers implemented using trigger volumes
Trigger Volume class for different terrain types can be placed on the map.
The terrain volumes can be scaled to cover large or small areas if desired,
but make sure to use snapping such that the borders are tight and there
is no overlap.

Movement speed for each terrain type is stored in a map(dictionary) with
the terrain classes as keys and speed modifiers as values.
This replaces the multiple individual speed modifier variables.
It is now very easy to add more Terrain types if desired.
A corresponding map can be used for terrain dependent damage modifiers.

The Unit class implements TerrainGetter component. This component can be
used by other classes as well, like the Tokens, for finding the terrain.

Units check for new terrain type with Event ActorEndOverlap when leaving
a terrain volume.
Terrain type defaults to Plains if a Unit is not touching any terrain volume.

Units now include their current terrain in their radio report.