major refactoring of Token-Unit pairing and player/enemy faction system
See issue #11 "Refactor Token-Unit pairing" on GitHub for a list of what
was done.

Now there are no specific classes for the Units and Tokens belonging to
the player and Enemy, so the Blueprints are much more tidy.

Instead you must set each Token's faction in the 'faction' variable in
each instance you place on the map. Faction 0=player, 1=enemy.

You must now explicitly link each Token to the corresponding Units that
you place on the board. The Unit's faction will automatically set to the
same faction as its Token, so there is no need to set it.
OR, you can avoid assigning a Unit to a Token. Then the Token the
correct type of Unit will automatically be created at the Token's
position.